+++ .idea/workspace.xml
@@ -7,13 +7,7 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="952c8f52-4fbc-4a9a-a04b-64cd54717eb6" name="Changes" comment="A little bug crept in to the chorus tag on songs.&#10;Displayed lyrics/custom slide text now respects empty lines.">
-      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ToDo" beforeDir="false" afterPath="$PROJECT_DIR$/ToDo" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/main.py" beforeDir="false" afterPath="$PROJECT_DIR$/main.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/settings_widget.py" beforeDir="false" afterPath="$PROJECT_DIR$/settings_widget.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/widgets.py" beforeDir="false" afterPath="$PROJECT_DIR$/widgets.py" afterDir="false" />
-    </list>
+    <list default="true" id="952c8f52-4fbc-4a9a-a04b-64cd54717eb6" name="Changes" comment="When selecting a custom font color, ensured that current color is default in the color dialog&#10;Ensured that cancelling the custom font color dialog doesn't change the current color to black&#10;Fixed delete image/background in settings widget to use the same delete function&#10;Ensured image widgets are updated after delete or add&#10;Changed the saving of settings to be threaded to facilitate more often saving of settings" />
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
@@ -168,14 +162,6 @@
       <workItem from="1725650345914" duration="5458000" />
       <workItem from="1725983928582" duration="6657000" />
       <workItem from="1725995303841" duration="13965000" />
-    </task>
-    <task id="LOCAL-00054" summary="Changed FontWidget's WindowType to a popup to facilitate auto-hide when clicked outside">
-      <option name="closed" value="true" />
-      <created>1725655471965</created>
-      <option name="number" value="00054" />
-      <option name="presentableId" value="LOCAL-00054" />
-      <option name="project" value="LOCAL" />
-      <updated>1725655471965</updated>
     </task>
     <task id="LOCAL-00055" summary="Changed FontWidget's WindowType to a popup to facilitate auto-hide when clicked outside">
       <option name="closed" value="true" />
@@ -561,7 +547,15 @@
       <option name="project" value="LOCAL" />
       <updated>1736366386874</updated>
     </task>
-    <option name="localTasksCounter" value="103" />
+    <task id="LOCAL-00103" summary="When selecting a custom font color, ensured that current color is default in the color dialog&#10;Ensured that cancelling the custom font color dialog doesn't change the current color to black&#10;Fixed delete image/background in settings widget to use the same delete function&#10;Ensured image widgets are updated after delete or add&#10;Changed the saving of settings to be threaded to facilitate more often saving of settings">
+      <option name="closed" value="true" />
+      <created>1736545879043</created>
+      <option name="number" value="00103" />
+      <option name="presentableId" value="LOCAL-00103" />
+      <option name="project" value="LOCAL" />
+      <updated>1736545879044</updated>
+    </task>
+    <option name="localTasksCounter" value="104" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -582,7 +576,6 @@
     </option>
   </component>
   <component name="VcsManagerConfiguration">
-    <MESSAGE value="Fixed default advanced options for new custom slide" />
     <MESSAGE value="Added missing add to service actions (image, video, and web) in the media_widget.py CustomListWidget" />
     <MESSAGE value="Added restore from backup option and corresponding code." />
     <MESSAGE value="Returned to bold being the default display widget lyric widget font style" />
@@ -607,7 +600,8 @@
     <MESSAGE value="Improved text handling in the edit widget's QTextEdit" />
     <MESSAGE value="PyQt6 broke audio and video. Reverted to PyQt5.&#10;Added coloring to verse tags in the edit widget." />
     <MESSAGE value="A little bug crept in to the chorus tag on songs.&#10;Displayed lyrics/custom slide text now respects empty lines." />
-    <option name="LAST_COMMIT_MESSAGE" value="A little bug crept in to the chorus tag on songs.&#10;Displayed lyrics/custom slide text now respects empty lines." />
+    <MESSAGE value="When selecting a custom font color, ensured that current color is default in the color dialog&#10;Ensured that cancelling the custom font color dialog doesn't change the current color to black&#10;Fixed delete image/background in settings widget to use the same delete function&#10;Ensured image widgets are updated after delete or add&#10;Changed the saving of settings to be threaded to facilitate more often saving of settings" />
+    <option name="LAST_COMMIT_MESSAGE" value="When selecting a custom font color, ensured that current color is default in the color dialog&#10;Ensured that cancelling the custom font color dialog doesn't change the current color to black&#10;Fixed delete image/background in settings widget to use the same delete function&#10;Ensured image widgets are updated after delete or add&#10;Changed the saving of settings to be threaded to facilitate more often saving of settings" />
   </component>
   <component name="XDebuggerManager">
     <watches-manager>
